# src/main/java/com/drip/competitionengine/api/dto/ReasonErrorDto.java
package com.drip.competitionengine.api.dto;
import java.time.Instant;
import java.util.UUID;
import lombok.*;

@Data @AllArgsConstructor @NoArgsConstructor
public class ReasonErrorDto { String reason; }